+++ changelog.md
@@ -5,9 +5,15 @@ All notable changes to the "QB64 PE" extension will be documented in this file.
 - Enhancements
   - Modify`tasks.json` to support QB64 PE.
 
+## 0.10.4
+- Enhancements
+  - Fixed new markdown hover help for QB64 PE keywords up to version `3.13.1`
+    - Now works with `config.helpPath`
+  - For how the markdown is generated look here: https://github.com/grymmjack/qb64pe-wiki-to-markdown
+
 ## 0.10.3
 - Enhancements
-  - Add new markdown hover helpf or QB64 PE keywords up to version `3.13.1`
+  - Add new markdown hover help for QB64 PE keywords up to version `3.13.1`
   - Modified wiki search online help to work with qb64phoenix.com/wiki/
 
 ## 0.10.2